♻️  Make minor improvements

+++ src/os/restart.sh
@@ -1,3 +1,21 @@
 #!/bin/bash
 
-sudo shutdown -r now &> /dev/null
+cd "$(dirname "${BASH_SOURCE[0]}")" \
+    && . "utils.sh"
+
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+main() {
+
+    print_in_purple "\n\n * Restart\n\n"
+
+    ask_for_confirmation "Do you want to restart?"
+    printf "\n"
+
+    if answer_is_yes; then
+        sudo shutdown -r now &> /dev/null
+    fi
+
+ }
+
+ main
